+++ src/main/liquibase/changelogs/homework-2/sql/07.create-order_item-table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE order_item (
-     order_id integer,
-     item_id integer,
+     order_id bigint,
+     item_id bigint,
      PRIMARY KEY (order_id, item_id)
 );
