+++ src/main/liquibase/changelogs/homework-2/sql/15.rent_order-add-customer_id.sql
@@ -1,2 +1,2 @@
 ALTER TABLE rent_order
-ADD COLUMN customer_id integer NOT NULL;
+ADD COLUMN customer_id bigint NOT NULL;
