+++ src/main/liquibase/changelogs/homework-2/sql/16.rent_order-add-worker_id.sql
@@ -1,2 +1,2 @@
 ALTER TABLE rent_order
-ADD COLUMN worker_id integer NOT NULL;
+ADD COLUMN worker_id bigint NOT NULL;
